backend/requirements.txt
django>=6.0.4
djangorestframework>=3.17.1
django-cors-headers>=4.9.0
pyjwt>=2.12.1
psycopg2-binary>=2.9.11
dj-database-url>=3.1.2
gunicorn>=23.0.0
whitenoise>=6.12.0
python-dotenv>=1.2.2
requests

